+++ src/main/java/com/utn/space/venueaapi/controllers/SpaceController.java
@@ -23,12 +23,12 @@ public class SpaceController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Space> findSpaceById(@PathVariable Long id){
+    public ResponseEntity<Space> findSpaceById(@PathVariable Integer id){
         return ResponseEntity.ok(spaceService.findById(id));
     }
 
     @DeleteMapping("/{id}")
-    public void deleteSpaceById(@PathVariable Long id){
+    public void deleteSpaceById(@PathVariable Integer id){
         spaceService.deleteById(id);
     }
 
@@ -38,7 +38,7 @@ public class SpaceController {
     }
 
     @PutMapping("/{id}")
-    public void modifySpace(@PathVariable Long id, @RequestBody SpaceDTO spaceDTO){
+    public void modifySpace(@PathVariable Integer id, @RequestBody SpaceDTO spaceDTO){
         spaceService.modifySpace(id,spaceDTO);
     }
