+++ src/main/java/com/utn/space/venueaapi/model/records/SpaceFilterDTO.java
@@ -1,8 +1,8 @@
 package com.utn.space.venueaapi.model.records;
 
 public record SpaceFilterDTO(
-        Long id_consumer_owner,
-        Long id_location,
+        Integer id_consumer_owner,
+        Integer id_location,
         Long name_space,
         Double minPrice,
         Double maxPrice)
