+++ src/main/java/com/utn/space/venueaapi/service/ServiceSelectedService.java
@@ -27,7 +27,7 @@ public class ServiceSelectedService {
         this.spaceServiceItemRepository = spaceServiceItemRepository;
     }
 
-    public List<ServiceSelectedDTO> getServicesSelectedOfReservation(Long idReservation){
+    public List<ServiceSelectedDTO> getServicesSelectedOfReservation(Integer idReservation){
         return repository.findServiceSelectedByIdReservation(idReservation);
     }
 
@@ -44,7 +44,7 @@ public class ServiceSelectedService {
     }
 
     @Transactional
-    public void insertListOfServicesSelectedInAReservation(Long idReservation, List<ServiceSelectedWithoutReservationDTO> servicesSelectedDTO){
+    public void insertListOfServicesSelectedInAReservation(Integer idReservation, List<ServiceSelectedWithoutReservationDTO> servicesSelectedDTO){
         Reservation reservation = reservationRepository.findById(idReservation).orElseThrow(() -> new NotFoundException("No se ha encontrado la reserva cuyo servicio quiere seleccionarse"));
         for(ServiceSelectedWithoutReservationDTO serviceSelectedDTO : servicesSelectedDTO){
             SpaceServiceItem serviceItem = spaceServiceItemRepository.findById(serviceSelectedDTO.idService()).orElseThrow(() -> new NotFoundException("No se ha encontrado el servicio que se quiere seleccionar en una reserva"));
@@ -55,7 +55,7 @@ public class ServiceSelectedService {
     }
 
     @Transactional
-    public void deleteServiceSelectedForAReservation(Long id){
+    public void deleteServiceSelectedForAReservation(Integer id){
         if(!repository.existsById(id)) throw new NotFoundException("No se encontró el servicio seleccionado para eliminarlo");
 
         repository.deleteById(id);
